requirements.txt
google-api-python-client==2.190.0
google-auth-httplib2==0.3.0
google-auth-oauthlib==1.2.4
pystray==0.19.5
Pillow==12.1.1
tzdata==2025.3
